fixes review/2
