+++ lib/barcodes_list_view.dart
@@ -1,7 +1,7 @@
-import 'package:barcode_widget/barcode_widget.dart';
 import 'package:flutter/material.dart';
 import 'barcode_item.dart';
 import 'barcode_detail_page.dart';
+import 'barcode_item_widget.dart'; 
 
 class BarcodesListView extends StatelessWidget {
   const BarcodesListView({
@@ -23,76 +23,43 @@ class BarcodesListView extends StatelessWidget {
 
         return Padding(
           padding: const EdgeInsets.symmetric(vertical: 4.0),
-          child: Row(
-            children: [
-              Expanded(
-                child: GestureDetector(
-                  onTap: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => BarcodeDetailPage(barcode: leftBarcode),
-                      ),
-                    );
-                  },
-                  child: BarcodeItemWidget(barcode: leftBarcode),
-                ),
-              ),
-              const SizedBox(width: 10),
-              if (rightBarcode != null)
-                Expanded(
-                  child: GestureDetector(
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => BarcodeDetailPage(barcode: rightBarcode),
-                        ),
-                      );
-                    },
-                    child: BarcodeItemWidget(barcode: rightBarcode),
-                  ),
-                ),
-            ],
-          ),
+          child: _buildBarcodeRow(context, leftBarcode, rightBarcode),
         );
       },
     );
   }
-}
 
-class BarcodeItemWidget extends StatelessWidget {
-  const BarcodeItemWidget({
-    super.key,
-    required this.barcode,
-  });
-
-  final BarcodeItem barcode;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
+  Widget _buildBarcodeRow(BuildContext context, BarcodeItem leftBarcode, BarcodeItem? rightBarcode) {
+    return Row(
       children: [
-        Text(
-          barcode.description,
-          style: const TextStyle(
-            fontSize: 17,
-            fontWeight: FontWeight.bold,
+        Expanded(
+          child: GestureDetector(
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => BarcodeDetailPage(barcode: leftBarcode),
+                ),
+              );
+            },
+            child: BarcodeItemWidget(barcode: leftBarcode),
           ),
         ),
-        const SizedBox(height: 6),
-        BarcodeWidget(
-          barcode: barcode.type,
-          data: barcode.data,
-          width: 200,
-          height: 100,
-          drawText: true,
-          style: const TextStyle(
-            fontSize: 16,
-            letterSpacing: 1.5,
+        const SizedBox(width: 10),
+        if (rightBarcode != null)
+          Expanded(
+            child: GestureDetector(
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => BarcodeDetailPage(barcode: rightBarcode),
+                  ),
+                );
+              },
+              child: BarcodeItemWidget(barcode: rightBarcode),
+            ),
           ),
-        ),
       ],
     );
   }
